elastic fix and retrieve book from personal book info fix

+++ app/service/elasticsearch/ElasticSearchClient.php
@@ -90,7 +90,7 @@ class ElasticSearchClient
                             ],
                             'retailPrice' => [
                                 'properties' => [
-                                    'amount' => [ 'type' => 'integer' ],
+                                    'amount' => [ 'type' => 'double' ],
                                     'currency' => ['type'=>'string']
                                 ]
                             ],
